moved enums to EapiFunctions.ts
